scatterplot data cleanup
