+++ src/stories/plots/ScatterPlot.storyData.ts
@@ -1,253 +1,535 @@
 import { NumberOrDateRange } from '../../types/general';
 import { ScatterPlotData } from '../../types/plots';
+import { min, max, lte, gte } from 'lodash';
+import {
+  gradientSequentialColorscaleMap,
+  gradientDivergingColorscaleMap,
+} from '../../types/plots/addOns';
+import { scaleLinear } from 'd3-scale';
 
 // set data array types for VEuPathDB scatter plot: https://redmine.apidb.org/issues/41310
 // but changed to new format: most likely x & y data are row/column vector format; also standardError is not a single value but vector
-interface VEuPathDBScatterPlotData<T extends number | Date> {
-  data: Array<{
-    series: {
-      x: T[]; // perhaps string[] is better despite Date format, e.g., ISO format?
-      y: T[]; // will y data have a Date?
-      popupContent?: string;
-    };
-    interval?: {
-      x: T[]; // perhaps string[] is better despite Date format, e.g., ISO format?
-      y: T[]; // will y data have a Date?
-      orientation: string;
-      standardError: number[];
-    };
-    color?: string;
-    label: string;
-    // for general scatter component
-    showLines?: boolean;
-    showMarkers?: boolean;
-    fillArea?: boolean;
-    useSpline?: boolean;
-  }>;
-  opacity?: number;
+export interface VEuPathDBScatterPlotData {
+  scatterplot: {
+    data: Array<{
+      seriesX?: number[] | string[]; // perhaps string[] is better despite Date format, e.g., ISO format?
+      seriesY?: number[] | string[]; // will y data have a Date?
+      smoothedMeanX?: number[] | string[]; // perhaps string[] is better despite Date format, e.g., ISO format?
+      smoothedMeanY?: number[]; // will y data have a date string? Nope, number only
+      smoothedMeanSE?: number[];
+      bestFitLineX?: number[] | string[];
+      bestFitLineY?: number[];
+      seriesGradientColorscale?: number[] | string[];
+    }>;
+  };
 }
 
-// change HTML hex code to rgb array
-const hexToRgb = (hex?: string): [number, number, number] => {
-  if (!hex) return [0, 0, 0];
-  const fullHex = hex.replace(
-    /^#?([a-f\d])([a-f\d])([a-f\d])$/i,
-    (m: string, r: string, g: string, b: string): string =>
-      '#' + r + r + g + g + b + b
-  );
-  const hexDigits = fullHex.substring(1);
-  const matches = hexDigits.match(/.{2}/g);
-  if (matches == null) return [0, 0, 0];
-  return matches.map((x: string) => parseInt(x, 16)) as [
-    number,
-    number,
-    number
-  ];
-};
-
-// check number array and if empty
-function isArrayOfNumbers(value: any): value is number[] {
-  // value.length !==0
-  return (
-    Array.isArray(value) &&
-    value.length !== 0 &&
-    value.every((item) => typeof item === 'number')
-  );
-}
-
-// an example data: data are assumed to be number type only
-let orientationValue = 'y';
-
-// Real data comprised of numbers
-const dataSet: VEuPathDBScatterPlotData<number> = {
-  data: [
-    {
-      // scatter plot with CI
-      series: {
-        x: [
-          0,
-          1,
-          2,
-          3,
-          4,
-          5,
-          6,
-          7,
-          8,
-          9,
-          10,
-          11,
-          12,
-          13,
-          14,
-          15,
-          16,
-          17,
-          18,
-          19,
-          20,
-          21,
-          22,
-          23,
-          24,
-          25,
-          26,
-          27,
-          28,
-          29,
-          30,
-          31,
-          32,
-          33,
-          34,
-          35,
-          36,
-          37,
-          38,
-          39,
-          40,
-          41,
-          42,
-          43,
-          44,
-          46,
-          47,
+// data with 'seriesGradientColorscale' column (for Gradient Colorscale). Sequential (all positive)
+export const dataSetSequentialGradient: VEuPathDBScatterPlotData = {
+  scatterplot: {
+    data: [
+      {
+        // scatter plot with gradient colorscale
+        seriesX: [
+          '0',
+          '1',
+          '2',
+          '3',
+          '4',
+          '5',
+          '6',
+          '7',
+          '8',
+          '9',
+          '10',
+          '11',
+          '12',
+          '13',
+          '14',
+          '15',
+          '16',
+          '17',
+          '18',
+          '19',
+          '20',
+          '21',
+          '22',
+          '23',
+          '24',
+          '25',
+          '26',
+          '27',
+          '28',
+          '29',
+          '30',
+          '31',
+          '32',
+          '33',
+          '34',
+          '35',
+          '36',
+          '37',
+          '38',
+          '39',
+          '40',
+          '41',
+          '42',
+          '43',
+          '44',
+          '46',
+          '47',
         ],
-        y: [
-          0.58,
-          0.54,
-          0.43,
-          0.86,
-          1.19,
-          1.47,
-          0.98,
-          1.36,
-          0.58,
-          0.82,
-          0.77,
-          1.22,
-          2.21,
-          0.46,
-          1.55,
-          1.38,
-          0.98,
-          1.4,
-          1.29,
-          1.3,
-          1.56,
-          1.73,
-          1.48,
-          1.38,
-          1.1,
-          1.14,
-          0.84,
-          1.12,
-          1.07,
-          1.1,
-          0.73,
-          0.86,
-          1.16,
-          1.02,
-          0.77,
-          1.04,
-          0.57,
-          0.08,
-          0.2,
-          0.4,
-          0.23,
-          0.13,
-          -0.51,
-          0,
-          -0.35,
-          0.21,
-          -0.08,
+        seriesY: [
+          '0.58',
+          '0.54',
+          '0.43',
+          '0.86',
+          '1.19',
+          '1.47',
+          '0.98',
+          '1.36',
+          '0.58',
+          '0.82',
+          '0.77',
+          '1.22',
+          '2.21',
+          '0.46',
+          '1.55',
+          '1.38',
+          '0.98',
+          '1.4',
+          '1.29',
+          '1.3',
+          '1.56',
+          '1.73',
+          '1.48',
+          '1.38',
+          '1.1',
+          '1.14',
+          '0.84',
+          '1.12',
+          '1.07',
+          '1.1',
+          '0.73',
+          '0.86',
+          '1.16',
+          '1.02',
+          '0.77',
+          '1.04',
+          '0.57',
+          '0.08',
+          '0.2',
+          '0.4',
+          '0.23',
+          '0.13',
+          '-0.51',
+          '0',
+          '-0.35',
+          '0.21',
+          '-0.08',
+        ],
+        // variable mapped to color (same as seriesX for easier checking the colorscale)
+        seriesGradientColorscale: [
+          '0',
+          '1',
+          '2',
+          '3',
+          '4',
+          '5',
+          '6',
+          '7',
+          '8',
+          '9',
+          '10',
+          '11',
+          '12',
+          '13',
+          '14',
+          '15',
+          '16',
+          '17',
+          '18',
+          '19',
+          '20',
+          '21',
+          '22',
+          '23',
+          '24',
+          '25',
+          '26',
+          '27',
+          '28',
+          '29',
+          '30',
+          '31',
+          '32',
+          '33',
+          '34',
+          '35',
+          '36',
+          '37',
+          '38',
+          '39',
+          '40',
+          '41',
+          '42',
+          '43',
+          '44',
+          '46',
+          '47',
         ],
       },
-      // popupContent?: string;   // this prop is placed at each object, {x, y, popupContent}, but not tested
-      interval: {
-        x: [
-          0,
-          0.5949367,
-          1.1898734,
-          1.7848101,
-          2.3797468,
-          2.9746835,
-          3.5696203,
-          4.164557,
-          4.7594937,
-          5.3544304,
-          5.9493671,
-          6.5443038,
-          7.1392405,
-          7.7341772,
-          8.3291139,
-          8.9240506,
-          9.5189873,
-          10.1139241,
-          10.7088608,
-          11.3037975,
-          11.8987342,
-          12.4936709,
-          13.0886076,
-          13.6835443,
-          14.278481,
-          14.8734177,
-          15.4683544,
-          16.0632911,
-          16.6582278,
-          17.2531646,
-          17.8481013,
-          18.443038,
-          19.0379747,
-          19.6329114,
-          20.2278481,
-          20.8227848,
-          21.4177215,
-          22.0126582,
-          22.6075949,
-          23.2025316,
-          23.7974684,
-          24.3924051,
-          24.9873418,
-          25.5822785,
-          26.1772152,
-          26.7721519,
-          27.3670886,
-          27.9620253,
-          28.556962,
-          29.1518987,
-          29.7468354,
-          30.3417722,
-          30.9367089,
-          31.5316456,
-          32.1265823,
-          32.721519,
-          33.3164557,
-          33.9113924,
-          34.5063291,
-          35.1012658,
-          35.6962025,
-          36.2911392,
-          36.8860759,
-          37.4810127,
-          38.0759494,
-          38.6708861,
-          39.2658228,
-          39.8607595,
-          40.4556962,
-          41.0506329,
-          41.6455696,
-          42.2405063,
-          42.835443,
-          43.4303797,
-          44.0253165,
-          44.6202532,
-          45.2151899,
-          45.8101266,
-          46.4050633,
-          47,
+    ],
+  },
+};
+
+// data with 'seriesGradientColorscale' column (for Gradient Colorscale). Diverging (some negative and some positive)
+export const dataSetDivergingGradient: VEuPathDBScatterPlotData = {
+  scatterplot: {
+    data: [
+      {
+        // scatter plot with gradient colorscale
+        seriesX: [
+          '0',
+          '1',
+          '2',
+          '3',
+          '4',
+          '5',
+          '6',
+          '7',
+          '8',
+          '9',
+          '10',
+          '11',
+          '12',
+          '13',
+          '14',
+          '15',
+          '16',
+          '17',
+          '18',
+          '19',
+          '20',
+          '21',
+          '22',
+          '23',
+          '24',
+          '25',
+          '26',
+          '27',
+          '28',
+          '29',
+          '30',
+          '31',
+          '32',
+          '33',
+          '34',
+          '35',
+          '36',
+          '37',
+          '38',
+          '39',
+          '40',
+          '41',
+          '42',
+          '43',
+          '44',
+          '46',
+          '47',
         ],
-        y: [
+        seriesY: [
+          '0.58',
+          '0.54',
+          '0.43',
+          '0.86',
+          '1.19',
+          '1.47',
+          '0.98',
+          '1.36',
+          '0.58',
+          '0.82',
+          '0.77',
+          '1.22',
+          '2.21',
+          '0.46',
+          '1.55',
+          '1.38',
+          '0.98',
+          '1.4',
+          '1.29',
+          '1.3',
+          '1.56',
+          '1.73',
+          '1.48',
+          '1.38',
+          '1.1',
+          '1.14',
+          '0.84',
+          '1.12',
+          '1.07',
+          '1.1',
+          '0.73',
+          '0.86',
+          '1.16',
+          '1.02',
+          '0.77',
+          '1.04',
+          '0.57',
+          '0.08',
+          '0.2',
+          '0.4',
+          '0.23',
+          '0.13',
+          '-0.51',
+          '0',
+          '-0.35',
+          '0.21',
+          '-0.08',
+        ],
+        // variable mapped to color (same as seriesX for easier checking the colorscale)
+        seriesGradientColorscale: [
+          '0.58',
+          '0.54',
+          '0.43',
+          '0.86',
+          '1.19',
+          '1.47',
+          '0.98',
+          '1.36',
+          '0.58',
+          '0.82',
+          '0.77',
+          '1.22',
+          '2.21',
+          '0.46',
+          '1.55',
+          '1.38',
+          '0.98',
+          '1.4',
+          '1.29',
+          '1.3',
+          '1.56',
+          '1.73',
+          '1.48',
+          '1.38',
+          '1.1',
+          '1.14',
+          '0.84',
+          '1.12',
+          '1.07',
+          '1.1',
+          '0.73',
+          '0.86',
+          '1.16',
+          '1.02',
+          '0.77',
+          '1.04',
+          '0.57',
+          '0.08',
+          '0.2',
+          '0.4',
+          '0.23',
+          '0.13',
+          '-0.51',
+          '0',
+          '-0.35',
+          '0.21',
+          '-0.08',
+        ],
+      },
+    ],
+  },
+};
+
+// use actual data response format (number/string) following scatter plot data API
+export const dataSet: VEuPathDBScatterPlotData = {
+  scatterplot: {
+    data: [
+      {
+        // scatter plot with CI
+        seriesX: [
+          '0',
+          '1',
+          '2',
+          '3',
+          '4',
+          '5',
+          '6',
+          '7',
+          '8',
+          '9',
+          '10',
+          '11',
+          '12',
+          '13',
+          '14',
+          '15',
+          '16',
+          '17',
+          '18',
+          '19',
+          '20',
+          '21',
+          '22',
+          '23',
+          '24',
+          '25',
+          '26',
+          '27',
+          '28',
+          '29',
+          '30',
+          '31',
+          '32',
+          '33',
+          '34',
+          '35',
+          '36',
+          '37',
+          '38',
+          '39',
+          '40',
+          '41',
+          '42',
+          '43',
+          '44',
+          '46',
+          '47',
+        ],
+        seriesY: [
+          '0.58',
+          '0.54',
+          '0.43',
+          '0.86',
+          '1.19',
+          '1.47',
+          '0.98',
+          '1.36',
+          '0.58',
+          '0.82',
+          '0.77',
+          '1.22',
+          '2.21',
+          '0.46',
+          '1.55',
+          '1.38',
+          '0.98',
+          '1.4',
+          '1.29',
+          '1.3',
+          '1.56',
+          '1.73',
+          '1.48',
+          '1.38',
+          '1.1',
+          '1.14',
+          '0.84',
+          '1.12',
+          '1.07',
+          '1.1',
+          '0.73',
+          '0.86',
+          '1.16',
+          '1.02',
+          '0.77',
+          '1.04',
+          '0.57',
+          '0.08',
+          '0.2',
+          '0.4',
+          '0.23',
+          '0.13',
+          '-0.51',
+          '0',
+          '-0.35',
+          '0.21',
+          '-0.08',
+        ],
+        // popupContent?: string;   // this prop is placed at each object, {x, y, popupContent}, but not tested
+        smoothedMeanX: [
+          '0',
+          '0.5949367',
+          '1.1898734',
+          '1.7848101',
+          '2.3797468',
+          '2.9746835',
+          '3.5696203',
+          '4.164557',
+          '4.7594937',
+          '5.3544304',
+          '5.9493671',
+          '6.5443038',
+          '7.1392405',
+          '7.7341772',
+          '8.3291139',
+          '8.9240506',
+          '9.5189873',
+          '10.1139241',
+          '10.7088608',
+          '11.3037975',
+          '11.8987342',
+          '12.4936709',
+          '13.0886076',
+          '13.6835443',
+          '14.278481',
+          '14.8734177',
+          '15.4683544',
+          '16.0632911',
+          '16.6582278',
+          '17.2531646',
+          '17.8481013',
+          '18.443038',
+          '19.0379747',
+          '19.6329114',
+          '20.2278481',
+          '20.8227848',
+          '21.4177215',
+          '22.0126582',
+          '22.6075949',
+          '23.2025316',
+          '23.7974684',
+          '24.3924051',
+          '24.9873418',
+          '25.5822785',
+          '26.1772152',
+          '26.7721519',
+          '27.3670886',
+          '27.9620253',
+          '28.556962',
+          '29.1518987',
+          '29.7468354',
+          '30.3417722',
+          '30.9367089',
+          '31.5316456',
+          '32.1265823',
+          '32.721519',
+          '33.3164557',
+          '33.9113924',
+          '34.5063291',
+          '35.1012658',
+          '35.6962025',
+          '36.2911392',
+          '36.8860759',
+          '37.4810127',
+          '38.0759494',
+          '38.6708861',
+          '39.2658228',
+          '39.8607595',
+          '40.4556962',
+          '41.0506329',
+          '41.6455696',
+          '42.2405063',
+          '42.835443',
+          '43.4303797',
+          '44.0253165',
+          '44.6202532',
+          '45.2151899',
+          '45.8101266',
+          '46.4050633',
+          '47',
+        ],
+        smoothedMeanY: [
           0.5785742,
           0.62672442,
           0.67323618,
@@ -329,7 +611,7 @@ const dataSet: VEuPathDBScatterPlotData<number> = {
           -0.25792504,
           -0.30664913,
         ],
-        standardError: [
+        smoothedMeanSE: [
           0.17707248,
           0.16235568,
           0.14874309,
@@ -411,409 +693,465 @@ const dataSet: VEuPathDBScatterPlotData<number> = {
           0.17443995,
           0.18947155,
         ],
-        orientation: orientationValue,
       },
-      color: '#00b0f6', // use this as fitting line and scatter color
-      label: 'Data 1', // use this as "name" in plotly
-      // below setting is for marker without line but use spline for fitting line
-      showLines: false,
-      showMarkers: true,
-      fillArea: false,
-      useSpline: true,
-    },
-    {
-      // line plot with marker
-      series: {
-        // random data from 0 to 29
-        x: [
-          0,
-          1,
-          2,
-          3,
-          4,
-          5,
-          6,
-          7,
-          8,
-          9,
-          10,
-          11,
-          12,
-          13,
-          14,
-          15,
-          16,
-          17,
-          18,
-          19,
-          20,
-          21,
-          22,
-          23,
-          24,
-          25,
-          26,
-          27,
-          28,
-          29,
+      {
+        // line plot with marker
+        seriesX: [
+          '0',
+          '1',
+          '2',
+          '3',
+          '4',
+          '5',
+          '6',
+          '7',
+          '8',
+          '9',
+          '10',
+          '11',
+          '12',
+          '13',
+          '14',
+          '15',
+          '16',
+          '17',
+          '18',
+          '19',
+          '20',
+          '21',
+          '22',
+          '23',
+          '24',
+          '25',
+          '26',
+          '27',
+          '28',
+          '29',
         ],
-        y: [
-          -1,
-          -0.8,
-          -0.4,
-          0,
-          0.5,
-          0.8,
-          1.3,
-          1.7,
-          2.0,
-          2.3,
-          2.7,
-          3.5,
-          4.0,
-          4.5,
-          4.7,
-          4.7,
-          4.6,
-          4.0,
-          3.5,
-          2.7,
-          2.3,
-          2.0,
-          1.7,
-          1.3,
-          0.8,
-          0.5,
-          0,
-          -0.4,
-          -0.67,
-          -1,
+        seriesY: [
+          '-1',
+          '-0.8',
+          '-0.4',
+          '0',
+          '0.5',
+          '0.8',
+          '1.3',
+          '1.7',
+          '2.0',
+          '2.3',
+          '2.7',
+          '3.5',
+          '4.0',
+          '4.5',
+          '4.7',
+          '4.7',
+          '4.6',
+          '4.0',
+          '3.5',
+          '2.7',
+          '2.3',
+          '2.0',
+          '1.7',
+          '1.3',
+          '0.8',
+          '0.5',
+          '0',
+          '-0.4',
+          '-0.67',
+          '-1',
         ],
       },
-      color: '#8B0000', // use this as point/marker color
-      label: 'Data 2', // use this as "name" in plotly
-      // below setting is for marker with line but no spline interpolation
-      showLines: true,
-      showMarkers: true,
-      fillArea: false,
-      useSpline: false,
-    },
-    {
-      // line plot with marker
-      series: {
-        // random data from 20 to 47
-        x: [
-          20,
-          21,
-          22,
-          23,
-          24,
-          25,
-          26,
-          27,
-          28,
-          29,
-          30,
-          31,
-          32,
-          33,
-          34,
-          35,
-          36,
-          37,
-          38,
-          39,
-          40,
-          41,
-          42,
-          43,
-          44,
-          45,
-          46,
-          47,
+      {
+        // line plot with marker
+        seriesX: [
+          '20',
+          '21',
+          '22',
+          '23',
+          '24',
+          '25',
+          '26',
+          '27',
+          '28',
+          '29',
+          '30',
+          '31',
+          '32',
+          '33',
+          '34',
+          '35',
+          '36',
+          '37',
+          '38',
+          '39',
+          '40',
+          '41',
+          '42',
+          '43',
+          '44',
+          '45',
+          '46',
+          '47',
         ],
-        y: [
-          -1,
-          -0.8,
-          -0.7,
-          -0.5,
-          -0.2,
-          0,
-          0.5,
-          0.8,
-          1.2,
-          1.5,
-          2.3,
-          2.6,
-          2.9,
-          3.0,
-          2.85,
-          2.6,
-          2.0,
-          1.5,
-          1.2,
-          1.0,
-          0.8,
-          0.4,
-          0,
-          -0.3,
-          -0.6,
-          -0.7,
-          -0.8,
-          -1,
+        seriesY: [
+          '-1',
+          '-0.8',
+          '-0.7',
+          '-0.5',
+          '-0.2',
+          '0',
+          '0.5',
+          '0.8',
+          '1.2',
+          '1.5',
+          '2.3',
+          '2.6',
+          '2.9',
+          '3.0',
+          '2.85',
+          '2.6',
+          '2.0',
+          '1.5',
+          '1.2',
+          '1.0',
+          '0.8',
+          '0.4',
+          '0',
+          '-0.3',
+          '-0.6',
+          '-0.7',
+          '-0.8',
+          '-1',
         ],
       },
-      color: '#006400', // use this as point/marker color
-      label: 'Data 3', // use this as "name" in plotly
-      // density plot: also testing for no line/marker => line seems to be defailt in this case
-      showLines: false,
-      showMarkers: false,
-      fillArea: true,
-      useSpline: true,
-    },
-  ],
-  // opacity: 0.2,
+    ],
+  },
 };
 
-// // a sample data for testing Date format: ISO format only
-// const dataSet: VEuPathDBScatterPlotData = {
-//   data: [
-//     {
-//       series: {
-//         x: ['2020-01-11T00:00:00Z', '2019-01-01T00:00:00Z', '2018-01-01T00:00:00Z', '2017-12-20T00:00:00Z', '2016-07-18T00:00:00Z', '2015-07-31T00:00:00Z','2014-01-11T00:00:00Z', '2013-01-01T00:00:00Z', '2012-05-01T00:00:00Z', '2012-01-05T00:00:00Z'],
-//         y: [1,2,3,4,5,6,7,8,9,10],
-//       },
-//       // popupContent?: string;   // this prop is placed at each object, {x, y, popupContent}, but not tested
-//       interval: {
-//         x: ['2020-01-11T00:00:00Z', '2019-01-01T00:00:00Z', '2018-01-01T00:00:00Z', '2017-12-20T00:00:00Z', '2016-07-18T00:00:00Z', '2015-07-31T00:00:00Z','2014-01-11T00:00:00Z', '2013-01-01T00:00:00Z', '2012-05-01T00:00:00Z', '2012-01-05T00:00:00Z'],
-//         y: [1.1,2.2,3.3,4.4,5.5,6.6,7.7,8.8,9.2,9.5],
-//         standardError: [0.1,0.145,0.2,0.11,0.5,0.24,0.7,0.33,0.45,0.55],
-//         orientation: orientationValue,
-//       },
-//       color: '#00b0f6',       // use this as fitting line and scatter color
-//       label: 'Dataset 1',     // use this as "name" in plotly
-//     },
-//   ],
-//   // opacity: 0.2,
-// }
+// a sample data for testing Date format: ISO format only
+export const dateStringDataSet: VEuPathDBScatterPlotData = {
+  scatterplot: {
+    data: [
+      // Case 2-1) X: date string; Y: number
+      {
+        // not sorted X
+        seriesX: [
+          '2012-05-01',
+          '2013-01-01',
+          '2012-01-05',
+          '2014-01-11',
+          '2015-07-31',
+          '2016-07-18',
+          '2017-12-20',
+          '2018-01-01',
+          '2019-01-01',
+          '2020-01-11',
+        ],
+        seriesY: ['2', '3', '1', '4', '5', '6', '7', '8', '9', '10'],
+        // not sorted X
+        smoothedMeanX: [
+          '2012-08-01',
+          '2013-06-01',
+          '2014-01-11',
+          '2015-05-31',
+          '2012-03-05',
+          '2016-02-18',
+          '2017-10-20',
+          '2018-04-01',
+          '2019-07-01',
+          '2019-12-11',
+        ],
+        smoothedMeanY: [2.2, 3.3, 4.4, 5.5, 1.1, 6.6, 7.7, 8.8, 9.2, 9.5],
+        smoothedMeanSE: [
+          0.145,
+          0.2,
+          0.11,
+          0.5,
+          0.1,
+          0.24,
+          0.7,
+          0.33,
+          0.45,
+          0.55,
+        ],
+      },
 
-// set global Opacity value
-let globalOpacity = dataSet.opacity ? dataSet.opacity : 1;
-
-// set a default color - set to any for now
-let defaultColor: string = '#00b0f6';
+      // // Case 2-2) X: number string; Y: date string
+      // // // In this case, smoothedMean and bestfitline have response error from backend anyway
+      // {
+      //   // not sorted X
+      //   seriesY: ['2012-05-01', '2013-01-01', '2012-01-05', '2014-01-11', '2015-07-31', '2016-07-18', '2017-12-20', '2018-01-01', '2019-01-01','2020-01-11' ],
+      //   seriesX: ["2","3","1","4","5","6","7","8","9","10"],
+      //   // // not sorted X
+      //   // smoothedMeanX: ['2012-08-01', '2013-06-01', '2014-01-11', '2015-05-31', '2012-03-05','2016-02-18', '2017-10-20', '2018-04-01', '2019-07-01', '2019-12-11' ],
+      //   // smoothedMeanY: [2.2,3.3,4.4,5.5,1.1,6.6,7.7,8.8,9.2,9.5],
+      //   // smoothedMeanSE: [0.145,0.2,0.11,0.5,0.1,0.24,0.7,0.33,0.45,0.55],
+      // },
+    ],
+  },
+};
 
 // making plotly input data
-function processInputData<T extends number | Date>(
-  dataSet: VEuPathDBScatterPlotData<T>
-): {
-  dataSetProcess: ScatterPlotData;
-  xAxisRange: NumberOrDateRange;
-  yAxisRange: NumberOrDateRange;
-} {
+export function processInputData<T extends number | string>(
+  dataSet: any,
+  vizType: string,
+  // line, marker,
+  modeValue: string,
+  // send independentValueType & dependentValueType
+  independentValueType: string,
+  dependentValueType: string,
+  defineColors: boolean
+) {
+  // set fillAreaValue for densityplot
+  const fillAreaValue = vizType === 'densityplot' ? 'toself' : '';
+
+  // distinguish data per Viztype
+  // currently, lineplot returning scatterplot, not lineplot
+  const plotDataSet =
+    vizType === 'lineplot'
+      ? dataSet.scatterplot
+      : vizType === 'densityplot'
+      ? // set densityplot as scatterplot for this example
+        // ? dataSet.densityplot
+        dataSet.scatterplot
+      : dataSet.scatterplot;
+
   // set variables for x- and yaxis ranges
-  let xMin: number | Date = 0;
-  let xMax: number | Date = 0;
-  let yMin: number | Date = 0;
-  let yMax: number | Date = 0;
+  let yMin: number | string | undefined = 0;
+  let yMax: number | string | undefined = 0;
 
-  let processedDataSeries: any = [];
-  dataSet.data.forEach(function (el: any, index: number) {
-    // initialize variables: setting with union type for future, but this causes typescript issue in the current version
-    let xSeriesValue: T[] = [];
-    let ySeriesValue: T[] = [];
-    let xIntervalLineValue: T[] = [];
-    let yIntervalLineValue: T[] = [];
-    let standardErrorValue: T[] = []; // this is for standardError
-    let xIntervalBounds: T[] = [];
-    let yIntervalBounds: T[] = [];
+  // coloring: using plotly.js default colors instead of web-components default palette
+  const markerColors = [
+    '31, 119, 180', //'#1f77b4',  // muted blue
+    '255, 127, 14', //'#ff7f0e',  // safety orange
+    '44, 160, 44', //'#2ca02c',  // cooked asparagus green
+    '214, 39, 40', //'#d62728',  // brick red
+    '148, 103, 189', //'#9467bd',  // muted purple
+    '140, 86, 75', //'#8c564b',  // chestnut brown
+    '227, 119, 194', //'#e377c2',  // raspberry yogurt pink
+    '127, 127, 127', //'#7f7f7f',  // middle gray
+    '188, 189, 34', //'#bcbd22',  // curry yellow-green
+    '23, 190, 207', //'#17becf'   // blue-teal
+  ];
 
-    // set rgbValue here per dataset with a default color
-    let rgbValue: number[] = el.color
-      ? hexToRgb(el.color)
-      : hexToRgb(defaultColor);
-    let scatterPointColor: string = '';
-    let fittingLineColor: string = '';
-    let intervalColor: string = '';
-    // set line and marker variable
-    let modeValue: string = '';
-    let splineValue: string = '';
-    let fillAreaValue: string = '';
+  // set dataSetProcess as any
+  let dataSetProcess: any = [];
+
+  // drawing raw data (markers) at first
+  plotDataSet.data.forEach(function (el: any, index: number) {
+    // initialize seriesX/Y
+    let seriesX = [];
+    let seriesY = [];
+
+    // initialize gradient colorscale arrays
+    let seriesGradientColorscale = [];
+    let markerColorsGradient = [];
 
     // series is for scatter plot
-    if (el.series) {
+    if (el.seriesX && el.seriesY) {
       // check the number of x = number of y
-      if (el.series.x.length !== el.series.y.length) {
-        console.log(
-          'x length=',
-          el.series.x.length,
-          '  y length=',
-          el.series.y.length
-        );
-        alert('The number of X data is not equal to the number of Y data');
+      if (el.seriesX.length !== el.seriesY.length) {
+        // alert('The number of X data is not equal to the number of Y data');
         throw new Error(
           'The number of X data is not equal to the number of Y data'
         );
       }
 
-      // probably no need to have this for series data, though
-      //1) combine the arrays:
-      let combinedArray = [];
-      for (let j = 0; j < el.series.x.length; j++) {
-        combinedArray.push({ xValue: el.series.x[j], yValue: el.series.y[j] });
-      }
-      //2) sort:
-      combinedArray.sort(function (a, b) {
-        return a.xValue < b.xValue ? -1 : a.xValue == b.xValue ? 0 : 1;
-      });
-      //3) separate them back out:
-      for (let k = 0; k < combinedArray.length; k++) {
-        xSeriesValue[k] = combinedArray[k].xValue;
-        ySeriesValue[k] = combinedArray[k].yValue;
-      }
-
       /*
-       *  set variables for x-/y-axes ranges including x,y data points: considering Date data for X as well
-       * This is for finding global min/max values among data arrays for better display of the plot(s)
-       */
-      // check if this X array consists of numbers & add type assertion
-      if (isArrayOfNumbers(xSeriesValue)) {
-        xMin =
-          xMin < Math.min(...(xSeriesValue as number[]))
-            ? xMin
-            : Math.min(...(xSeriesValue as number[]));
-        xMax =
-          xMax > Math.max(...(xSeriesValue as number[]))
-            ? xMax
-            : Math.max(...(xSeriesValue as number[]));
+        For raw data, there are two cases:
+          a) X: number string; Y: date string
+          b) X: date string; Y: number string
+        For the case of b), smoothed mean and best fit line option would get backend response error
+      **/
+      if (independentValueType === 'date') {
+        seriesX = el.seriesX;
       } else {
-        // this array consists of Dates
-        if (index == 0) {
-          // to set initial min/max Date values for Date[]
-          xMin = getMinDate(xSeriesValue as Date[]);
-          xMax = getMaxDate(xSeriesValue as Date[]);
-        } else {
-          xMin =
-            xMin <
-            Math.min(...xSeriesValue.map((date) => new Date(date).getTime()))
-              ? xMin
-              : new Date(
-                  Math.min(
-                    ...xSeriesValue.map((date) => new Date(date).getTime())
-                  )
-                );
-          xMax =
-            xMax >
-            Math.max(...xSeriesValue.map((date) => new Date(date).getTime()))
-              ? xMax
-              : new Date(
-                  Math.max(
-                    ...xSeriesValue.map((date) => new Date(date).getTime())
-                  )
-                );
-        }
+        seriesX = el.seriesX.map(Number);
+      }
+      if (dependentValueType === 'date') {
+        seriesY = el.seriesY;
+      } else {
+        seriesY = el.seriesY.map(Number);
       }
 
       // check if this Y array consists of numbers & add type assertion
-      if (isArrayOfNumbers(ySeriesValue)) {
-        if (index == 0) {
-          yMin = Math.min(...ySeriesValue);
-          yMax = Math.max(...ySeriesValue);
-        } else {
-          yMin =
-            yMin < Math.min(...ySeriesValue) ? yMin : Math.min(...ySeriesValue);
-          yMax =
-            yMax > Math.max(...ySeriesValue) ? yMax : Math.max(...ySeriesValue);
-        }
+      if (index === 0) {
+        // if (seriesY && seriesY !== undefined) {
+        yMin = min(seriesY);
+        yMax = max(seriesY);
+        // }
       } else {
-        if (index == 0) {
-          // to set initial Date value for Date[]
-          yMin = getMinDate(ySeriesValue as Date[]);
-          yMax = getMaxDate(ySeriesValue as Date[]);
+        yMin =
+          // (yMin !== undefined && seriesY.length !== 0 && yMin < min(seriesY)) ? yMin : min(seriesY);
+          lte(yMin, min(seriesY)) ? yMin : min(seriesY);
+        yMax = gte(yMax, max(seriesY)) ? yMax : max(seriesY);
+      }
+
+      // If seriesGradientColorscale column exists, need to use gradient colorscales
+      if (el.seriesGradientColorscale) {
+        // Assuming only allowing numbers for now - later will add dates
+        seriesGradientColorscale = el.seriesGradientColorscale.map(Number);
+
+        // For storybook only - determine overlayMin, overlayMax, and gradientColorscaleType inside processInputData
+        // In web-eda, these need to be determined *outside* of this function so that we handle facets correctly.
+        let overlayMin: number | undefined;
+        let overlayMax: number | undefined;
+        let gradientColorscaleType: string | undefined;
+
+        const defaultOverlayMin: number = min(
+          seriesGradientColorscale
+        ) as number;
+        const defaultOverlayMax: number = max(
+          seriesGradientColorscale
+        ) as number;
+        // Note overlayMin and/or overlayMax could be intentionally 0.
+        gradientColorscaleType =
+          defaultOverlayMin >= 0 && defaultOverlayMax >= 0
+            ? 'sequential'
+            : defaultOverlayMin <= 0 && defaultOverlayMax <= 0
+            ? 'sequential reversed'
+            : 'divergent';
+
+        // Update overlay min and max
+        if (gradientColorscaleType === 'divergent') {
+          overlayMin = -Math.max(
+            Math.abs(defaultOverlayMin),
+            Math.abs(defaultOverlayMax)
+          );
+          overlayMax = Math.max(
+            Math.abs(defaultOverlayMin),
+            Math.abs(defaultOverlayMax)
+          );
         } else {
-          yMin =
-            yMin < getMinDate(ySeriesValue as Date[])
-              ? yMin
-              : getMinDate(ySeriesValue as Date[]);
-          yMax =
-            yMax > getMaxDate(ySeriesValue as Date[])
-              ? yMax
-              : getMaxDate(ySeriesValue as Date[]);
+          overlayMin = defaultOverlayMin;
+          overlayMax = defaultOverlayMax;
         }
-      }
+        // -- end for storybook only code section
 
-      // use global opacity for coloring
-      scatterPointColor =
-        'rgba(' +
-        rgbValue[0] +
-        ',' +
-        rgbValue[1] +
-        ',' +
-        rgbValue[2] +
-        ',' +
-        globalOpacity +
-        ')'; // set alpha/opacity as 0.2 for CI
+        // Determine marker colors
+        if (
+          gradientColorscaleType &&
+          (overlayMin || overlayMin === 0) &&
+          overlayMax
+        ) {
+          // If we have data, use a gradient colorscale. No data series will have all NaN values in seriesGradientColorscale
 
-      // check plot options: default value at plotly js seems to be lines
-      if (el.showLines == true && el.showMarkers == true) {
-        modeValue = 'lines+markers';
-      } else if (el.showLines == true && el.showMarkers == false) {
-        modeValue = 'lines';
-      } else if (el.showLines == false && el.showMarkers == true) {
-        modeValue = 'markers';
-      }
+          // Initialize normalization function.
+          const normalize = scaleLinear();
 
-      if (el.useSpline) {
-        splineValue = 'spline';
-      }
+          if (gradientColorscaleType === 'divergent') {
+            // Diverging colorscale, assume 0 is midpoint. Colorscale must be symmetric around the midpoint
+            const maxAbsOverlay =
+              Math.abs(overlayMin) > overlayMax
+                ? Math.abs(overlayMin)
+                : overlayMax;
 
-      if (el.fillArea) {
-        fillAreaValue = 'toself';
+            // For each point, normalize the data to [-1, 1], then retrieve the corresponding color
+            normalize.domain([-maxAbsOverlay, maxAbsOverlay]).range([-1, 1]);
+            markerColorsGradient = seriesGradientColorscale.map((a: number) =>
+              gradientDivergingColorscaleMap(normalize(a))
+            );
+          } else if (gradientColorscaleType === 'sequntial reverse') {
+            // Normalize data to [1, 0], so that the colorscale goes in reverse. NOTE: can remove once we add the ability for users to set colorscale range.
+            normalize.domain([overlayMin, overlayMax]).range([1, 0]);
+            markerColorsGradient = seriesGradientColorscale.map((a: number) =>
+              gradientSequentialColorscaleMap(normalize(a))
+            );
+            gradientColorscaleType = 'sequential';
+          } else {
+            // Then we use the sequential (from 0 to inf) colorscale.
+            // For each point, normalize the data to [0, 1], then retrieve the corresponding color
+            normalize.domain([overlayMin, overlayMax]).range([0, 1]);
+            markerColorsGradient = seriesGradientColorscale.map((a: number) =>
+              gradientSequentialColorscaleMap(normalize(a))
+            );
+            gradientColorscaleType = 'sequential';
+          }
+        }
       }
 
       // add scatter data considering input options
-      processedDataSeries.push({
-        x: xSeriesValue,
-        y: ySeriesValue,
-        name: el.label,
-        // mode: 'markers',
-        // mode: 'lines+markers',
+      dataSetProcess.push({
+        x: seriesX,
+        y: seriesY,
+        // distinguish X/Y Data from Overlay
+        name: el.overlayVariableDetails
+          ? el.overlayVariableDetails.value
+          : 'Data',
         mode: modeValue,
-        // type: 'scattergl',
-        type: 'scatter',
+        type:
+          vizType === 'lineplot'
+            ? 'scatter'
+            : vizType === 'densityplot'
+            ? 'scatter'
+            : 'scattergl', // for the raw data of the scatterplot
         fill: fillAreaValue,
-        marker: { color: scatterPointColor, size: 12 },
-        line: { color: scatterPointColor, shape: splineValue },
+        marker: {
+          color: defineColors
+            ? 'rgb(' + markerColors[index] + ')'
+            : seriesGradientColorscale?.length > 0
+            ? markerColorsGradient
+            : undefined,
+          size: 12,
+          line: { color: 'rgb(' + markerColors[index] + ')', width: 2 },
+          symbol: 'circle',
+        },
+        // this needs to be here for the case of markers with line or lineplot.
+        // always use spline?
+        line: {
+          color: defineColors ? 'rgb(' + markerColors[index] + ')' : undefined,
+          shape: 'spline',
+        },
       });
     }
+  });
 
-    // check if interval prop exists
-    if (el.interval) {
+  // after drawing raw data, smoothedMean and bestfitline plots are displayed
+  plotDataSet.data.forEach(function (el: any, index: number) {
+    // initialize variables: setting with union type for future, but this causes typescript issue in the current version
+    let xIntervalLineValue: T[] = [];
+    let yIntervalLineValue: number[] = [];
+    let standardErrorValue: number[] = []; // this is for standardError
+
+    let xIntervalBounds: T[] = [];
+    let yIntervalBounds: number[] = [];
+
+    // initialize smoothedMeanX, bestFitLineX
+    let smoothedMeanX = [];
+    let bestFitLineX = [];
+
+    // check if smoothedMean prop exists
+    if (el.smoothedMeanX && el.smoothedMeanY && el.smoothedMeanSE) {
       // check the number of x = number of y or standardError
-      if (
-        el.interval.x.length !== el.interval.y.length ||
-        el.interval.x.length !== el.interval.standardError.length
-      ) {
-        alert(
-          'The number of X data is not equal to the number of Y data or standardError data'
-        );
+      if (el.smoothedMeanX.length !== el.smoothedMeanY.length) {
         throw new Error(
           'The number of X data is not equal to the number of Y data or standardError data'
         );
       }
+
+      // change string array to number array for numeric data
+      // smoothedMeanX = el.smoothedMeanX.map(Number);
+      if (independentValueType === 'date') {
+        smoothedMeanX = el.smoothedMeanX;
+      } else {
+        smoothedMeanX = el.smoothedMeanX.map(Number);
+      }
+      // smoothedMeanY/SE are number[]
+
+      // the date format, yyyy-mm-dd works with sort, so no change in the following is required
       // sorting function
       //1) combine the arrays: including standardError
       let combinedArrayInterval = [];
-      for (let j = 0; j < el.interval.x.length; j++) {
+      for (let j = 0; j < smoothedMeanX.length; j++) {
         combinedArrayInterval.push({
-          xValue: el.interval.x[j],
-          yValue: el.interval.y[j],
-          zValue: el.interval.standardError[j],
+          xValue: smoothedMeanX[j],
+          yValue: el.smoothedMeanY[j],
+          zValue: el.smoothedMeanSE[j],
         });
       }
       //2) sort:
       combinedArrayInterval.sort(function (a, b) {
-        return a.xValue < b.xValue ? -1 : a.xValue == b.xValue ? 0 : 1;
+        return a.xValue < b.xValue ? -1 : a.xValue === b.xValue ? 0 : 1;
       });
       //3) separate them back out:
       for (let k = 0; k < combinedArrayInterval.length; k++) {
@@ -822,77 +1160,38 @@ function processInputData<T extends number | Date>(
         standardErrorValue[k] = combinedArrayInterval[k].zValue;
       }
 
-      // set variables for x-/y-axes ranges including fitting line: note that initial values of xMin and xMax are already defined earlier (el.series)
-      if (isArrayOfNumbers(xIntervalLineValue)) {
-        xMin = el.series.x
-          ? xMin < Math.min(...xIntervalLineValue)
-            ? xMin
-            : Math.min(...xIntervalLineValue)
-          : Math.min(...xIntervalLineValue);
-        xMax = el.series.x
-          ? xMax > Math.max(...xIntervalLineValue)
-            ? xMax
-            : Math.max(...xIntervalLineValue)
-          : Math.max(...xIntervalLineValue);
-      } else {
-        xMin = el.series.x
-          ? xMin < getMinDate(xIntervalLineValue as Date[])
-            ? xMin
-            : getMinDate(xIntervalLineValue as Date[])
-          : getMinDate(xIntervalLineValue as Date[]);
-        xMax = el.series.x
-          ? xMax > getMaxDate(xIntervalLineValue as Date[])
-            ? xMax
-            : getMaxDate(xIntervalLineValue as Date[])
-          : getMaxDate(xIntervalLineValue as Date[]);
-      }
+      // add additional condition for the case of smoothedMean (without series data)
+      yMin = el.seriesY
+        ? lte(yMin, min(yIntervalLineValue))
+          ? yMin
+          : min(yIntervalLineValue)
+        : min(yIntervalLineValue);
+      yMax = el.seriesY
+        ? gte(yMax, max(yIntervalLineValue))
+          ? yMax
+          : max(yIntervalLineValue)
+        : max(yIntervalLineValue);
 
-      if (isArrayOfNumbers(yIntervalLineValue)) {
-        yMin = el.series.y
-          ? yMin < Math.min(...yIntervalLineValue)
-            ? yMin
-            : Math.min(...yIntervalLineValue)
-          : Math.min(...yIntervalLineValue);
-        yMax = el.series.y
-          ? yMax > Math.max(...yIntervalLineValue)
-            ? yMax
-            : Math.max(...yIntervalLineValue)
-          : Math.max(...yIntervalLineValue);
-      } else {
-        yMin = el.series.y
-          ? yMin < getMinDate(yIntervalLineValue as Date[])
-            ? yMin
-            : getMinDate(yIntervalLineValue as Date[])
-          : getMinDate(yIntervalLineValue as Date[]);
-        yMax = el.series.y
-          ? yMax > getMaxDate(yIntervalLineValue as Date[])
-            ? yMax
-            : getMaxDate(yIntervalLineValue as Date[])
-          : getMaxDate(yIntervalLineValue as Date[]);
-      }
-      // use global opacity for coloring
-      fittingLineColor =
-        'rgba(' +
-        rgbValue[0] +
-        ',' +
-        rgbValue[1] +
-        ',' +
-        rgbValue[2] +
-        ',' +
-        globalOpacity +
-        ')';
-
-      // store data for fitting line: this is not affected by plot options (e.g., showLine etc.)
-      processedDataSeries.push({
+      // store data for smoothed mean: this is not affected by plot options (e.g., showLine etc.)
+      dataSetProcess.push({
         x: xIntervalLineValue,
         y: yIntervalLineValue,
-        name: el.label + ' fitting',
-        // mode: 'lines+markers',
+        // name: 'Smoothed mean',
+        name: el.overlayVariableDetails
+          ? el.overlayVariableDetails.value + '<br>Smoothed mean'
+          : 'Smoothed mean',
         mode: 'lines', // no data point is displayed: only line
-        // type: 'line',
-        // line: {color: el.color, shape: 'spline',  width: 5 },
-        line: { color: fittingLineColor, shape: 'spline', width: 5 },
+        line: {
+          color: defineColors
+            ? 'rgba(' + markerColors[index] + ',1)'
+            : undefined,
+          shape: 'spline',
+          width: 2,
+        },
+        // use scattergl
+        type: 'scattergl',
       });
+
       // make Confidence Interval (CI) or Bounds (filled area)
       xIntervalBounds = xIntervalLineValue;
       xIntervalBounds = xIntervalBounds.concat(
@@ -901,9 +1200,7 @@ function processInputData<T extends number | Date>(
 
       // finding upper and lower bound values.
       const { yUpperValues, yLowerValues } = getBounds(
-        el.interval.orientation === 'x'
-          ? xIntervalLineValue
-          : yIntervalLineValue,
+        yIntervalLineValue,
         standardErrorValue
       );
 
@@ -913,53 +1210,100 @@ function processInputData<T extends number | Date>(
         yLowerValues.map((element: any) => element).reverse()
       );
 
-      // set alpha/opacity as 0.2 for CI
-      intervalColor =
-        'rgba(' + rgbValue[0] + ',' + rgbValue[1] + ',' + rgbValue[2] + ',0.2)';
-
       // set variables for x-/y-axes ranges including CI/bounds: no need for x data as it was compared before
-      yMin =
-        yMin < Math.min(...yLowerValues.map(Number))
-          ? yMin
-          : Math.min(...yLowerValues.map(Number));
-      yMax =
-        yMax > Math.max(...yUpperValues.map(Number))
-          ? yMax
-          : Math.max(...yUpperValues.map(Number));
+      yMin = lte(yMin, min(yLowerValues)) ? yMin : min(yLowerValues);
+      yMax = gte(yMax, max(yLowerValues)) ? yMax : max(yLowerValues);
 
       // store data for CI/bounds
-      processedDataSeries.push({
+      dataSetProcess.push({
         x: xIntervalBounds,
         y: yIntervalBounds,
-        name: 'Confidence interval',
-        fill: 'tozerox',
-        fillcolor: intervalColor,
-        type: 'line',
-        line: { color: 'transparent', shape: 'spline' }, // here, line means upper and lower bounds
+        // name: '95% Confidence interval',
+        name: el.overlayVariableDetails
+          ? el.overlayVariableDetails.value + '<br>95% Confidence interval'
+          : '95% Confidence interval',
+        // this is better to be tozeroy, not tozerox
+        fill: 'tozeroy',
+        fillcolor: defineColors
+          ? 'rgba(' + markerColors[index] + ',0.2)'
+          : undefined,
+        // type: 'line',
+        type: 'scattergl',
+        // here, line means upper and lower bounds
+        line: {
+          color: defineColors ? 'transparent' : undefined,
+          shape: 'spline',
+        },
       });
     }
 
-    // make some margin for y-axis range (5% of range for now)
-    if (typeof yMin == 'number' && typeof yMax == 'number') {
-      yMin = yMin - (yMax - yMin) * 0.05;
-      yMax = yMax + (yMax - yMin) * 0.05;
+    // accomodating bestFitLineWithRaw
+    // check if bestFitLineX/Y props exist
+    if (el.bestFitLineX && el.bestFitLineY) {
+      // check the number of x = number of y
+      if (el.bestFitLineX.length !== el.bestFitLineY.length) {
+        throw new Error(
+          'The number of X data is not equal to the number of Y data or standardError data'
+        );
+      }
+
+      // change string array to number array for numeric data
+      if (independentValueType === 'date') {
+        bestFitLineX = el.bestFitLineX;
+      } else {
+        bestFitLineX = el.bestFitLineX.map(Number);
+      }
+
+      // add additional condition for the case of smoothedMean (without series data)
+      yMin = el.seriesY
+        ? lte(yMin, min(el.bestFitLineY))
+          ? yMin
+          : min(el.bestFitLineY)
+        : min(el.bestFitLineY);
+      yMax = el.seriesY
+        ? gte(yMax, max(el.bestFitLineY))
+          ? yMax
+          : max(el.bestFitLineY)
+        : max(el.bestFitLineY);
+
+      // store data for fitting line: this is not affected by plot options (e.g., showLine etc.)
+      dataSetProcess.push({
+        x: bestFitLineX,
+        y: el.bestFitLineY,
+        // display R-square value at legend text(s)
+        // name: 'Best fit<br>R<sup>2</sup> = ' + el.r2,
+        name: el.overlayVariableDetails
+          ? el.overlayVariableDetails.value + '<br>R<sup>2</sup> = ' + el.r2
+          : 'Best fit<br>R<sup>2</sup> = ' + el.r2,
+        mode: 'lines', // no data point is displayed: only line
+        line: {
+          color: defineColors
+            ? 'rgba(' + markerColors[index] + ',1)'
+            : undefined,
+          shape: 'spline',
+        },
+        // use scattergl
+        type: 'scattergl',
+      });
     }
   });
 
-  return {
-    dataSetProcess: { series: processedDataSeries },
-    xAxisRange: { min: xMin, max: xMax },
-    yAxisRange: { min: yMin, max: yMax },
-  };
+  // make some margin for y-axis range (5% of range for now)
+  if (typeof yMin == 'number' && typeof yMax == 'number') {
+    yMin = yMin - (yMax - yMin) * 0.05;
+    yMax = yMax + (yMax - yMin) * 0.05;
+  } else {
+    // set yMin/yMax to be NaN so that plotly uses autoscale for date type
+    yMin = NaN;
+    yMax = NaN;
+  }
+
+  return { dataSetProcess: { series: dataSetProcess }, yMin, yMax };
 }
 
-function getMinDate(dates: Date[]) {
-  return new Date(Math.min(...dates.map(Number)));
-}
-
-function getMaxDate(dates: Date[]) {
-  return new Date(Math.max(...dates.map(Number)));
-}
+/*
+ * Utility functions for processInputData()
+ */
 
 function getBounds<T extends number | Date>(
   values: T[],
@@ -979,7 +1323,3 @@ function getBounds<T extends number | Date>(
 
   return { yUpperValues, yLowerValues };
 }
-
-export const { dataSetProcess, xAxisRange, yAxisRange } = processInputData(
-  dataSet
-);
